+++ src/app/page.tsx
@@ -6,7 +6,7 @@ export default function Home() {
         <Container h="95vh" size="xl" mt="xl">
             <Stack h="100%" gap="xs">
                 <Title order={2} ta="center">
-                    To-do list!!
+                    To-do list
                 </Title>
                 <TaskBoard />
             </Stack>