+++ src/components/TaskColumn.tsx
@@ -1,6 +1,13 @@
 "use client";
 
-import { Group, Paper, ScrollAreaAutosize, Stack, Title } from "@mantine/core";
+import {
+    Grid,
+    Group,
+    Paper,
+    ScrollAreaAutosize,
+    Stack,
+    Title,
+} from "@mantine/core";
 import { TaskCard } from "./TaskCard/TaskCard";
 import { useDroppable } from "@dnd-kit/react";
 import { ColumnId, Task } from "types/board.types";
@@ -25,43 +32,59 @@ export const TaskColumn = memo(
         handleAddTask,
         handleContentChange,
     }: TaskColumnProps) => {
+        const isNote = id === "NOTE";
+        const type = isNote ? DND_TYPES.NOTE_COLUMN : DND_TYPES.TASK_COLUMN;
+        const accept = isNote ? DND_TYPES.NOTE_ITEM : DND_TYPES.TASK_ITEM;
+
         const { ref } = useDroppable({
             id,
-            type: DND_TYPES.TASK_COLUMN,
-            accept: DND_TYPES.TASK_ITEM,
+            type,
+            accept,
         });
 
         return (
             <Paper
                 ref={ref}
                 h="100%"
+                mih={0}
                 p="md"
-                bg="violet.0"
+                bg="themeColor.1"
                 shadow="xs"
                 withBorder
             >
-                <Stack h="100%" gap="sm">
-                    <Group justify="space-between">
-                        <Title order={4} fw="bold">
-                            {label}
-                        </Title>
-                        {id === "IN_PROGRESS" && (
-                            <NewTaskButton onClick={() => handleAddTask(id)}>
-                                <RiAddLine />
-                            </NewTaskButton>
-                        )}
-                    </Group>
+                <Stack h="100%" miw={0} mih={0} gap="sm">
+                    <Grid align="center">
+                        <Grid.Col span={4} />
+
+                        <Grid.Col span={4}>
+                            <Title ta="center" order={4} fw="bold">
+                                {label}
+                            </Title>
+                        </Grid.Col>
+
+                        <Grid.Col span={4}>
+                            <Group justify="flex-end">
+                                {(id === "IN_PROGRESS" || id === "NOTE") && (
+                                    <NewTaskButton
+                                        onClick={() => handleAddTask(id)}
+                                    >
+                                        <RiAddLine />
+                                    </NewTaskButton>
+                                )}
+                            </Group>
+                        </Grid.Col>
+                    </Grid>
 
                     <ScrollAreaAutosize
-                        h="100%"
-                        mah={745}
+                        flex={1}
                         type="scroll"
                         scrollbarSize={8}
                     >
-                        <Stack gap="xs">
+                        <Stack gap="sm">
                             {tasks.map((task, index) => (
                                 <TaskCard
                                     key={task.id}
+                                    isNote={isNote}
                                     task={task}
                                     index={index}
                                     handleContentChange={handleContentChange}